requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.3
django-paypal==2.1
idna==3.10
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
